+++ kernel/kernel-defaults/src/main/java/io/delta/kernel/defaults/internal/expressions/ElementAtEvaluator.java
@@ -16,7 +16,7 @@
 package io.delta.kernel.defaults.internal.expressions;
 
 import static io.delta.kernel.defaults.internal.DefaultEngineErrors.unsupportedExpressionException;
-import static io.delta.kernel.defaults.internal.expressions.ImplicitCastExpression.canCastTo;
+import static io.delta.kernel.defaults.internal.expressions.CastExpressionEvaluator.canCastTo;
 import static io.delta.kernel.internal.util.Preconditions.checkArgument;
 import static java.lang.String.format;
 
@@ -50,7 +50,7 @@ class ElementAtEvaluator {
 
     if (!keyTypeFromMapInput.equivalent(lookupKeyType)) {
       if (canCastTo(lookupKeyType, keyTypeFromMapInput)) {
-        lookupKey = new ImplicitCastExpression(lookupKey, keyTypeFromMapInput);
+        lookupKey = new CastExpressionEvaluator(lookupKey, keyTypeFromMapInput);
       } else {
         String reason =
             format(
